Docs: DESIGN: Move punctuation outside quotes
In particular:

  * It's called "stupidsum", not "stupidsum,".

  * It's called an "exponential distribution", not
    an "exponential distribution.".

  * It's called a "fanout", not a "fanout.".

  * That's "frequently", not "frequently.".

  * The term is "seekable", not "seekable.".

  * It's called "the index", not "the index.".

  * It's called "dirty", not "dirty,".

Just to drive home the point, consider these lines from the text:

  * the main program that runs when you type 'bup'.

  * you have several choices; the main ones are 'bup restore',
    'bup ftp', 'bup fuse', and 'bup web'.

  * You generate midx files with 'bup midx'.

  * it belongs in a tool like 'bup fsck' or 'git fsck'.

As you can see, it would be illogical and even harmful to include
the punctuation as part of the quotation; a period does NOT belong
in the command.

What if Avery had been overheard saying:

  I dislike bup for a lot of reasons, but it has the most potential
  of any backup tool yet created.

and then he were quoted in text with:

  Avery was overheard saying "I dislike bup for a lot of reasons."

To me, that is an immoral, unconscionable misquote; the quoted
period was not at all part of what Avery said, and it lends a
finality to the statement that was never there in the first
place. While it would be better to place that statement in more
context, it would at least be advisable to remove the period
from the quote (as it was never part of the quote in the first
place!):

  Avery was overheard saying "I dislike bup for a lot of reasons".

Of course, if the quoted sentence does indeed end right next
to the end of the sentence in which the quote is embedded,
then I think a good, logical case can be made for allowing
a `punctuational contraction'. For instance, rather than
ending this sentence with two periods nearby each other:

  Avery was overheard saying "I dislike bup for a lot of reasons,
  but it has the most potential of any backup tool yet created.".

it would seem appropriate to merge the periods into one period
within the quote:

  Avery was overheard saying "I dislike bup for a lot of reasons,
  but it has the most potential of any backup tool yet created."

Signed-off-by: Michael Witten <mfwitten@gmail.com>