Docs: DESIGN: Clean up the info about each of the lower order bits being 1
Firstly, I think the apostrophe in the original text:

  if they're all 1's

is not correct on a logical level. Nevertheless, it doesn't matter
now anyway, because the discussion has been converted to use the
singular form: `each bit' is discussed rather than `all bits' are
discussed.

Some wording was improved along the way.

Signed-off-by: Michael Witten <mfwitten@gmail.com>